pyproject.toml
[tool.poetry]
name = "leo"
version = "0.1.0"
description = ""
authors = ["Yeojin Kim <yeojin.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
pymongo = "^3.10.1"
python-dotenv = "^0.14.0"
pytz = "^2020.1"
slackclient = "^2.7.3"

[tool.poetry.dev-dependencies]
isort = "^5.0.9"
flake8 = "^3.8.3"
pytest = "^5.4.3"
coveralls = "^2.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
